flask-con-orm/requirements.txt
alembic==1.9.4
aniso8601==9.0.1
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
psycopg2-binary==2.9.5
python-dotenv==1.0.0
pytz==2022.7.1
six==1.16.0
SQLAlchemy==2.0.4
typing_extensions==4.5.0
Werkzeug==2.2.3
